Rename package `jobmanager.protocol' to `jobmanager.connector'.
