+++ workspace/popjava/src/popjava/service/jobmanager/POPJavaJobManager.java
@@ -39,9 +39,9 @@ import popjava.service.jobmanager.network.NodeJobManager;
 import popjava.service.jobmanager.network.POPNetwork;
 import popjava.service.jobmanager.network.POPNetworkNode;
 import popjava.service.jobmanager.network.POPNetworkNodeFactory;
-import popjava.service.jobmanager.protocol.POPConnectorBase;
-import popjava.service.jobmanager.protocol.POPConnectorFactory;
-import popjava.service.jobmanager.protocol.POPConnectorJobManager;
+import popjava.service.jobmanager.connector.POPConnectorBase;
+import popjava.service.jobmanager.connector.POPConnectorFactory;
+import popjava.service.jobmanager.connector.POPConnectorJobManager;
 import popjava.service.jobmanager.search.SNExploration;
 import popjava.service.jobmanager.search.SNNodesInfo;
 import popjava.service.jobmanager.search.SNRequest;
